requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.1
Flask-Modus==0.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
networkx==2.8.5
numpy==1.23.1
pycparser==2.21
PyMySQL==1.0.2
scikit-fuzzy==0.4.2
scipy==1.9.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
zipp==3.8.1
